Fixed multiple textures at once, now there can be 32 unique textures rendered at once.

+++ src/Graphics/Vertice.h
@@ -9,7 +9,7 @@ namespace Fireblast
 		glm::vec3 Vertice;
 		glm::vec4 Color;
 		glm::vec2 Uv;
-		unsigned int TextureID;
+		float TextureID;
 	};
 
 	struct Vertex3D
